prod: hardening sprint (auth, TLS, observability, durability, encryption, backup, CI, multi-node)
Phase A — Durability + observability
  * Fix `LocalFsStore::put` to actually fsync when durable=true (was a silent
    data-loss bug; the compactor's segment writes were never durabilized).
  * Default durable=true; opt out via `LocalFsStore::new_unsafe_fast` or
    `ZEN_UNSAFE_FAST=1`. Removes the prior `ZEN_FS_DURABLE=1`-to-opt-in landmine.
  * Add `metrics-exporter-prometheus` `/v1/metrics` endpoint with histograms +
    counters for every customer hot path. Wire OTLP tracing exporter from
    existing `telemetry.otlp_endpoint` config.
  * Group-commit fsync coordinator (`zen_storage::group_commit`): concurrent
    durable writers share one syscall via a broadcast barrier. Single-writer
    cost +1 µs; concurrent throughput recovered.

Phase B — Auth + TLS
  * New `zen_auth` crate: JWT (RS256 + JWKS) + claims cache (moka, 16K cap,
    5-min TTL); HMAC-SHA256 for inter-node `/v1/internal/*` with replay
    protection and constant-time compare.
  * Axum middleware: layered routers (public / customer / internal). 401 on
    missing/expired token; 401 on bad HMAC.
  * `RemoteClient` signs outbound inter-node requests with the same secret.
  * TLS termination via `axum-server` + `rustls 0.23.40` + `aws-lc-rs`. Cert
    paths in `cfg.server.tls.{cert_path, key_path}`. Optional — falls back
    to plaintext when unset (run behind a TLS-terminating LB).
  * 8 zen_auth lib tests + 4 auth integration tests + 1 TLS smoke test.

Phase C — CI/CD + zero clippy warnings
  * `.github/workflows/{ci,release}.yml`, dependabot, deny.toml, CODEOWNERS.
  * Drove 83 clippy warnings to 0; CI enforces `-D warnings`.

Phase D — Operational hardening
  * Migrations via `sqlx::migrate!("./migrations/sqlite")` with versioned
    SQL files (0001_initial_schema, 0002_nodes_registry).
  * Graceful shutdown on SIGTERM/SIGINT; flushes memtables to WAL before exit.
  * Per-tenant rate limit (`governor` token bucket) + global concurrency cap
    (`tower::limit::ConcurrencyLimitLayer`).
  * Health probes split: `/v1/healthz` (liveness) vs `/v1/readyz` (catalog
    probe).
  * Helm chart hardening: PodDisruptionBudget, HPA, NetworkPolicy,
    ServiceAccount with IRSA hooks, Ingress, separate liveness/readiness.
  * Terraform AWS module hardening: KMS key, SSE-KMS bucket, Secrets Manager
    for DB password, multi-AZ RDS with deletion protection + Performance
    Insights, least-privilege IAM, BYO VPC.

Phase E — Encryption at rest + backup/restore
  * New `zen_crypto` crate: per-segment AES-256-GCM envelope encryption
    with pluggable RootKey (StaticRootKey from disk; KMS adapter is the
    next step).
  * `EncryptingStore` BlobStore wrapper, opt-in via `cfg.crypto.enabled`
    (default off — existing benchmarks unaffected). Detects `ZENV` magic
    on read; legacy unencrypted blobs pass through.
  * `zen admin-backup` / `zen admin-restore` CLI commands serialize a
    tenant's segments + manifest to a directory and replay them against
    a fresh catalog. Integration test proves round-trip preserves rows.

Phase F — Tests + docs + runbook
  * Operator runbook (`docs/RUNBOOK.md`) covering deploy, observability,
    auth, backup/restore, scaling, common incidents, decommission.
  * `CHANGELOG.md`, README updates with prod-readiness section.
  * Chaos tests: kill_writer (WAL-replay recovery), disk_full (ENOSPC
    surfaces as structured error, no panic).
  * Property tests for WAL framing + ZenithQL parser + segment reader.
  * cargo-fuzz harnesses (separate workspace) for WAL/segment/QL parsers.

Bug fixes surfaced by property tests
  * `SegmentReader::from_bytes`: bounds-checked the metadata + hotcache
    slice ranges. Previously a corrupted footer offset could trigger an
    out-of-range panic — a DoS if an attacker controlled segment bytes.

Test growth: 137 → 182 (+45 tests). Workspace clippy warnings: 83 → 0.

+++ crates/zen_index/benches/index_bench.rs
@@ -39,7 +39,7 @@ fn bench_posting_intersect(c: &mut Criterion) {
     }
     let pa = PostingList { bitmap: a };
     let pb = PostingList { bitmap: b };
-    group.throughput(Throughput::Elements((pa.cardinality() + pb.cardinality()) as u64));
+    group.throughput(Throughput::Elements(pa.cardinality() + pb.cardinality()));
     group.bench_function("AND_1M_rows", |bb| {
         bb.iter(|| {
             let mut x = pa.clone();
